Unify text inputs
