+++ backend/src/packages/chaiNNer_external/external_stable_diffusion/automatic1111/txt2img.py
@@ -20,7 +20,7 @@ from nodes.properties.inputs import (
     EnumInput,
     SeedInput,
     SliderInput,
-    TextAreaInput,
+    TextInput,
 )
 from nodes.properties.outputs import ImageOutput
 from nodes.utils.seed import Seed
@@ -37,8 +37,8 @@ verify_api_connection()
     description="Generate an image using Automatic1111",
     icon="BsFillImageFill",
     inputs=[
-        TextAreaInput("Prompt", has_handle=True).make_optional(),
-        TextAreaInput("Negative Prompt", has_handle=True).make_optional(),
+        TextInput("Prompt", multiline=True).make_optional(),
+        TextInput("Negative Prompt", multiline=True).make_optional(),
         group("seed")(SeedInput()),
         SliderInput("Steps", minimum=1, default=20, maximum=150),
         EnumInput(
