+++ src/renderer/components/inputs/SchemaInput.tsx
@@ -22,7 +22,6 @@ import { HandleWrapper, InputContainer, WithLabel } from './InputContainer';
 import { NumberInput } from './NumberInput';
 import { InputProps } from './props';
 import { SliderInput } from './SliderInput';
-import { TextAreaInput } from './TextAreaInput';
 import { TextInput } from './TextInput';
 
 const InputComponents: {
@@ -33,8 +32,7 @@ const InputComponents: {
 } = {
     file: FileInput,
     directory: DirectoryInput,
-    'text-line': TextInput,
-    text: TextAreaInput,
+    text: TextInput,
     dropdown: DropDownInput,
     number: NumberInput,
     slider: SliderInput,
